chore: break off unassigned activities section of invoice form
